Bias evaluation metrics

Bias evaluation metrics

bias(
  group = "all",
  month = "all",
  err,
  model.name,
  keys,
  plot = FALSE,
  calibrate = "RAW"
)

Arguments

group
#demographic or patient groups, defaults at all

month
#temporal segmentation of the data, defaults at all -- i.e., no time stamps

err
##the error file generated from mlho learn

model.name
##name of the model

keys
##phenotype and model information for merging extracted from err

plot
##wanna plot out reliability diagrams?

calibrate
##wanna calibrate the computed probabilities?

Value